@@ -1,50 +1,59 @@
 {
-  "name": "chrome-extension-typescript-starter",
-  "version": "1.0.0",
-  "description": "chrome-extension-typescript-starter",
-  "main": "index.js",
-  "scripts": {
-    "watch": "webpack --config webpack/webpack.dev.js --watch",
-    "build": "webpack --config webpack/webpack.prod.js && npx tailwindcss -o dist/output.css --minify",
-    "clean": "rimraf dist",
-    "test": "npx jest",
-    "style": "prettier --write \"src/**/*.{ts,tsx}\""
-  },
-  "author": "",
+  "name": "chatgpt-enhancement-extension",
+  "version": "1.2.0",
+  "description": "ChatGPT Markdown and PDF Extension",
   "license": "MIT",
   "repository": {
     "type": "git",
-    "url": "https://github.com/chibat/chrome-extension-typescript-starter.git"
+    "url": "https://github.com/Jonghakseo/chrome-extension-boilerplate-react-vite.git"
   },
+  "scripts": {
+    "build": "tsc --noEmit && vite build",
+    "build:watch": "cross-env __DEV__=true vite build --watch",
+    "build:hmr": "rollup --config utils/reload/rollup.config.ts",
+    "wss": "node utils/reload/initReloadServer.js",
+    "dev": "npm run build:hmr && (run-p wss build:watch)",
+    "test": "jest"
+  },
+  "type": "module",
   "dependencies": {
-    "@emotion/react": "^11.10.6",
-    "@emotion/styled": "^11.10.6",
-    "@mui/icons-material": "^5.11.16",
-    "@mui/material": "^5.11.16",
-    "@mui/styled-engine-sc": "^5.11.11",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "styled-components": "^5.3.9"
+    "clsx": "^1.2.1",
+    "react": "18.2.0",
+    "react-dom": "18.2.0"
   },
   "devDependencies": {
-    "@types/chrome": "0.0.158",
-    "@types/jest": "^27.0.2",
-    "@types/react": "^18.0.29",
-    "@types/react-dom": "^18.0.11",
-    "copy-webpack-plugin": "^9.0.1",
-    "css-loader": "^6.7.3",
-    "glob": "^7.1.6",
-    "jest": "^27.2.1",
-    "pdfjs-dist": "^3.4.120",
-    "prettier": "^2.2.1",
-    "rimraf": "^3.0.2 ",
-    "style-loader": "^3.3.2",
-    "tailwindcss": "^3.3.0",
-    "ts-jest": "^27.0.5",
-    "ts-loader": "^8.0.0",
-    "typescript": "^4.4.3 ",
-    "webpack": "^5.76.0",
-    "webpack-cli": "^4.0.0",
-    "webpack-merge": "^5.0.0"
+    "@rollup/plugin-typescript": "^8.5.0",
+    "@testing-library/react": "13.4.0",
+    "@types/chrome": "0.0.224",
+    "@types/jest": "29.0.3",
+    "@types/node": "18.15.11",
+    "@types/react": "18.0.21",
+    "@types/react-dom": "18.0.11",
+    "@types/ws": "^8.5.3",
+    "@typescript-eslint/eslint-plugin": "5.56.0",
+    "@typescript-eslint/parser": "5.38.1",
+    "@vitejs/plugin-react": "2.2.0",
+    "autoprefixer": "^10.4.14",
+    "chokidar": "^3.5.3",
+    "cross-env": "^7.0.3",
+    "eslint": "8.36.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-prettier": "4.2.1",
+    "eslint-plugin-react": "7.32.2",
+    "fs-extra": "11.1.0",
+    "jest": "29.0.3",
+    "jest-environment-jsdom": "29.5.0",
+    "npm-run-all": "^4.1.5",
+    "pdfjs-dist": "^3.5.141",
+    "postcss": "^8.4.21",
+    "prettier": "2.7.1",
+    "rollup": "2.79.1",
+    "sass": "1.55.0",
+    "tailwindcss": "^3.3.1",
+    "ts-jest": "29.0.2",
+    "ts-loader": "9.4.2",
+    "typescript": "4.8.3",
+    "vite": "3.1.3",
+    "ws": "8.9.0"
   }
 }
